refactor(publish): split 'superdesk-publish' module.
`superdesk-publish` becomes:
```
.
├── controllers
│   ├── AdminPublishSettingsController.js
│   ├── DataConsistencyController.js
│   ├── PublishQueueController.js
│   └── index.js
├── directives
│   ├── DestinationDirective.js
│   ├── SubscribersDirective.js
│   └── index.js
├── services
│   ├── AdminPublishSettingsService.js
│   ├── SubscribersService.js
│   └── index.js
├── styles
│   └── publish.less
├── tests
│   └── publish.spec.js
├── views
│   ├── data-consistency.html
│   ├── destination.html
│   ├── email-config.html
│   ├── file-config.html
│   ├── ftp-config.html
│   ├── http-push-config.html
│   ├── odbc-config.html
│   ├── publish-queue.html
│   ├── settings.html
│   └── subscribers.html
├── constants.js
└── index.js

6 directories, 24 files
```
