+++ scripts/superdesk-publish/controllers/DataConsistencyController.js
@@ -1,0 +1,19 @@
+DataConsistencyController.$inject = ['$scope', 'api'];
+export function DataConsistencyController($scope, api) {
+    $scope.consistency_records = null;
+
+    function fetchConsistencyRecords () {
+        var criteria = criteria || {};
+        criteria.max_results = 200;
+        return api.consistency.query(criteria);
+    }
+
+    $scope.reload = function() {
+        fetchConsistencyRecords().then(function(data) {
+            $scope.consistency_records = data._items;
+            $scope.lastRefreshedAt = new Date();
+        });
+    };
+
+    $scope.reload ();
+}
