+++ scripts/superdesk-publish/controllers/index.js
@@ -1,0 +1,3 @@
+export { AdminPublishSettingsController } from './AdminPublishSettingsController';
+export { DataConsistencyController } from './DataConsistencyController';
+export { PublishQueueController } from './PublishQueueController';
