+++ scripts/superdesk-publish/directives/index.js
@@ -1,0 +1,2 @@
+export { DestinationDirective } from './DestinationDirective';
+export { SubscribersDirective } from './SubscribersDirective';
