+++ scripts/superdesk-publish/services/index.js
@@ -1,0 +1,2 @@
+export { SubscribersService } from './SubscribersService';
+export { AdminPublishSettingsService } from './AdminPublishSettingsService';
